Various changes towards supporting nested systems
- Moves pipe_channel out to its own files.
- Makes a new user_endpoint for users connecting into systems.
- Requires systems to be closed now with given external connections.
