+++ library/source/flow/file_endpoint.cpp
@@ -4,7 +4,9 @@ namespace flow {
 
 auto operator<<(std::ostream& os, const file_endpoint& value) -> std::ostream&
 {
+    os << "file_endpoint{";
     os << value.path;
+    os << "}";
     return os;
 }
